Migrate decoder FieldType methods into subpackage
